use setup.py entrypoints to install console script
Using custom console script breaks when the default python is not the
same python version as the python version that azure-cli was installed
for, i.e. if azure-cli is installed via pip3 and default python for the
system is python2.

Using setuptools entry_points/console_scripts feature sets up a console
script that specifically targets the python version that the package was
installed for.